+++ packages/wagmi/index.ts
@@ -1,108 +1,2 @@
-import { SafeRelayer, UserRelayerProps } from "@skaleboarder/safe-tools"
-import { providers, ethers, Signer } from "ethers"
-import { Connector } from "wagmi"
-
-type Address = string
-
-interface ContractDeploy {
-    address: Address
-}
-
-export interface WagmiWrapperConfig {
-    ethers: typeof ethers
-    provider: providers.Provider
-    chainId: string,
-    deploys: {
-        GnosisSafe: ContractDeploy,
-        GnosisSafeProxyFactory: ContractDeploy,
-        MultiSend: ContractDeploy,
-        MultiSendCallOnly: ContractDeploy,
-        CompatibilityFallbackHandler: ContractDeploy,
-        SignMessageLib: ContractDeploy,
-        CreateCall: ContractDeploy,
-        EnglishOwnerAdder: ContractDeploy,
-        WalletDeployer: ContractDeploy,
-    }
-    localStorage?: UserRelayerProps["localStorage"]
-}
-
-class WagmiWrapper {
-    private config: WagmiWrapperConfig
-
-    constructor(config: WagmiWrapperConfig) {
-        this.config = config
-    }
-
-    private contractConfigs() {
-        return {
-            [this.config.chainId]: {
-                safeMasterCopyAddress: this.config.deploys.GnosisSafe.address,
-                safeProxyFactoryAddress: this.config.deploys.GnosisSafeProxyFactory.address,
-                multiSendAddress: this.config.deploys.MultiSend.address,
-                multiSendCallOnlyAddress: this.config.deploys.MultiSendCallOnly.address,
-                fallbackHandlerAddress: this.config.deploys.CompatibilityFallbackHandler.address,
-                signMessageLibAddress: this.config.deploys.SignMessageLib.address,
-                createCallAddress: this.config.deploys.CreateCall.address,
-            }
-        }
-    }
-
-    private createRelayer = (signer: Signer) => {
-        return new SafeRelayer({
-            ethers,
-            signer,
-            EnglishOwnerAdderAddress: this.config.deploys.EnglishOwnerAdder.address,
-            walletDeployerAddress: this.config.deploys.WalletDeployer.address,
-            networkConfig: this.contractConfigs(),
-            provider: this.config.provider,
-            localStorage: this.config.localStorage,
-            faucet: async (address: string) => {
-                // TODO: demonstrate PoW on an schain
-                const resp = await fetch("/api/localFaucet", {
-                    method: "POST",
-                    body: JSON.stringify({ address })
-                })
-                console.log("faucet succeeded", address, await resp.json())
-            }
-        })
-    }
-
-    wrapConnector(connector: Connector) {
-        // console.log("wrapping", connector)
-        let signerPromise: Promise<Signer>
-
-        const proxy = new Proxy(connector, {
-            get: (target, prop) => {
-                switch (prop) {
-                    case "getProvider":
-                        // console.log('get provider: ', this.config.provider)
-                        return async () => {
-                            return this.config.provider
-                        }
-                    case "getSigner":
-                        console.log("get getSigner")
-                        return () => {
-                            if (signerPromise) {
-                                return signerPromise
-                            }
-                            signerPromise = (async () => {
-                                const original = await target.getSigner()
-                                const relayer = this.createRelayer(original)
-                                return relayer.wrappedSigner()
-                            })()
-                            return signerPromise
-                        }
-                    default:
-                        const original = (target as any)[prop]
-                        if (typeof original === "function") {
-                            return original.bind(target)
-                        }
-                        return original
-                }
-            }
-        })
-        return proxy
-    }
-}
-
-export default WagmiWrapper
+export * from "./src/wrapper"
+export * from "./src/chain"
